sorting based on value for my optimization

+++ src/Paintings.cpp
@@ -3,13 +3,13 @@
 //
 
 #include "Paintings.h"
-void Paintings::setWidth(int w) {width = w;}
+void Paintings::setWidth(double w) {width = w;}
 void Paintings::setHeight(int h) {height = h;}
-void Paintings::setPrice(int p) {price = p;}
+void Paintings::setPrice(double p) {price = p;}
 void Paintings::setID(int i) {id = i;}
 
-int Paintings::getWidth() {return width;}
+double Paintings::getWidth() {return width;}
 int Paintings::getHeight() {return height;}
-int Paintings::getPrice() {return price;}
+double Paintings::getPrice() {return price;}
 int Paintings::getID() {return id;}
 double Paintings::getValue() {return price/width;}
